lending_requirements.txt
pandas
matplotlib
sqlalchemy
mysql-connector-python
python-dotenv
